# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Finance MVP API
  version: 0.1.0
  summary: Canonical OpenAPI contract for the mounted closed-MVP runtime surface.
  license:
    name: Proprietary/Internal
    url: https://example.invalid/proprietary-internal-license
  description: |
    This document is the source of truth for the mounted MVP wire routes, DTO field names,
    enum values, error codes, pagination envelopes, and auth requirements.

    MVP boundaries are intentionally narrow: financial data is manual-entry only,
    `sourceType` accepts only `manual`, and transfers are same-scope only. Safe
    screenshot OCR capture may create pending drafts from sanitized structured
    candidates, but raw image bytes and OCR text are rejected and confirmed
    transactions remain `sourceType=manual`. There is no real import, parser,
    file storage, bank API, automated ingestion, broker credential, external
    credential, card, IBAN, raw statement, or device-token endpoint.

    Privacy invariants:
    - personal data is owner-only;
    - shared data is visible only to active members of the same Household;
    - reports support `personal`, `shared_family_report`, and `combined_viewer_overview`;
    - every list, report, search, autocomplete, cache, count, and aggregation
      filters visible rows before any aggregation or pagination;
    - responses never expose hidden counts, filtered-out counts, hidden facets,
      or diagnostics that confirm inaccessible resources.
servers:
...
security:
  - pwaCookieAuth: []
    csrfHeader: []
  - androidBearerAuth: []
paths:
...
  /planning/plans:
    get:
      tags: [Planning]
      operationId: getPlanningPlanForScopeMonth
      summary: Get plan for scope and month.
      description: Resolves personal owner-only or household active-membership scope before loading income sources, allocations, or totals. If month is omitted, the server uses the next calendar month.
      parameters:
        - $ref: '#/components/parameters/PlanningScopeRequired'
        - $ref: '#/components/parameters/PlanningMonth'
        - $ref: '#/components/parameters/HouseholdIdFilter'
      responses:
        '200':
          description: Planning plan for the requested scope/month.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningPlanEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Planning]
      operationId: createPlanningPlan
      summary: Create personal or household monthly plan.
      description: A plan has one currency and one scope. Duplicate scope/month plans return conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningPlanCreateRequest'
      responses:
        '201':
          description: Created planning plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningPlanEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/plans/history:
    get:
      tags: [Planning]
      operationId: listPlanningPlanHistory
      summary: List historical plans for a scope.
      description: Returns visible plan summaries sorted by month descending; no hidden counts are exposed.
      parameters:
        - $ref: '#/components/parameters/PlanningScopeRequired'
        - $ref: '#/components/parameters/HouseholdIdFilter'
      responses:
        '200':
          description: Visible planning history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningPlanSummaryListEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/plans/{planId}:
    get:
      tags: [Planning]
      operationId: getPlanningPlan
      summary: Get visible planning plan.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Planning plan detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningPlanEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/plans/{planId}/income-sources:
    post:
      tags: [Planning]
      operationId: createPlanningIncomeSource
      summary: Add planned income source.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningIncomeSourceCreateRequest'
      responses:
        '201':
          description: Created income source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningIncomeSourceEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/plans/{planId}/allocations:
    post:
      tags: [Planning]
      operationId: createPlanningAllocation
      summary: Add amount or percent allocation.
      description: Target must be a visible same-scope expense category, same-currency account, or same-currency account-backed asset/investment. Overallocated plans are allowed and surfaced through derived flags.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningAllocationCreateRequest'
      responses:
        '201':
          description: Created allocation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningAllocationEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/plans/{planId}/copy:
    post:
      tags: [Planning]
      operationId: copyPlanningPlan
      summary: Copy plan to another month.
      description: Copies income sources and allocations, preserving allocation mode/value. Missing or inaccessible targets become requiresAttention rows with targetSnapshot; no category/account/asset/transaction is auto-created.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningPlanCopyRequest'
      responses:
        '201':
          description: Copied planning plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningPlanEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/income-sources/{incomeSourceId}:
    patch:
      tags: [Planning]
      operationId: updatePlanningIncomeSource
      summary: Update planned income source.
      parameters:
        - $ref: '#/components/parameters/IncomeSourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningIncomeSourceUpdateRequest'
      responses:
        '200':
          description: Updated income source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningIncomeSourceEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Planning]
      operationId: deletePlanningIncomeSource
      summary: Delete planned income source.
      parameters:
        - $ref: '#/components/parameters/IncomeSourceId'
      responses:
        '204':
          description: Income source deleted.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/income-sources/{incomeSourceId}/confirm:
    post:
      tags: [Planning]
      operationId: confirmPlanningIncomeSource
      summary: Confirm income source in planning only.
      description: Sets confirmation state and audit-ish actor fields on the planning row only. It does not create transactions or mutate balances.
      parameters:
        - $ref: '#/components/parameters/IncomeSourceId'
      responses:
        '200':
          description: Confirmed income source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningIncomeSourceEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /planning/allocations/{allocationId}:
    patch:
      tags: [Planning]
      operationId: updatePlanningAllocation
      summary: Update allocation.
      parameters:
        - $ref: '#/components/parameters/AllocationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningAllocationUpdateRequest'
      responses:
        '200':
          description: Updated allocation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanningAllocationEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Planning]
      operationId: deletePlanningAllocation
      summary: Delete allocation.
      parameters:
        - $ref: '#/components/parameters/AllocationId'
      responses:
        '204':
          description: Allocation deleted.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    pwaCookieAuth:
      type: apiKey
      in: cookie
      name: __Host-finance_session
      description: HttpOnly Secure SameSite cookie used by the PWA.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for PWA cookie-authenticated state-changing requests; bound to the authenticated session and rotated on login/logout/reset.
    androidBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token for Android; refresh token storage and rotation are platform secure-storage concerns.
  responses:
    Error:
      description: Canonical error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  parameters:
...
    PlanId:
      name: planId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
    IncomeSourceId:
      name: incomeSourceId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
    AllocationId:
      name: allocationId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
    PlanningScopeRequired:
      name: scope
      in: query
      required: true
      schema: { $ref: '#/components/schemas/PlanningScope' }
    PlanningMonth:
      name: month
      in: query
      required: false
      schema: { $ref: '#/components/schemas/PlanMonth' }
      description: Calendar month accepted as YYYY-MM or YYYY-MM-01. Defaults to next month when omitted.
...
    HouseholdIdFilter:
      name: householdId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
      description: Optional scope narrowing only; inaccessible ids use neutral behavior.
...
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9_:-]+$'
      examples: [usr_123]
...
    PlanMonth:
      type: string
      pattern: '^[0-9]{4}-[0-9]{2}(-01)?$'
      examples: ['2026-06', '2026-06-01']
      description: Calendar month accepted as YYYY-MM or YYYY-MM-01. Responses use YYYY-MM; server stores the first day of the month.
...
    CurrencyCode:
      type: string
      pattern: '^[A-Z]{3}$'
      examples: [RUB]
...
    PlanningScope:
      type: string
      enum: [personal, household]
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDto'
...
    PlanningPlanDto:
      type: object
      additionalProperties: false
      required: [id, scope, month, currency, incomeSources, allocations, summary, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        scope:
          $ref: '#/components/schemas/PlanningScope'
        ownerUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        month:
          $ref: '#/components/schemas/PlanMonth'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        incomeSources:
          type: array
          items:
            $ref: '#/components/schemas/PlanningIncomeSourceDto'
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/PlanningAllocationDto'
        summary:
          $ref: '#/components/schemas/PlanningSummaryDto'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        version:
          type: integer
          minimum: 1
...
    PlanningPlanCreateRequest:
      type: object
      additionalProperties: false
      required: [scope, month, currency]
      properties:
        id:
          type: string
          format: uuid
        scope:
          $ref: '#/components/schemas/PlanningScope'
        month:
          $ref: '#/components/schemas/PlanMonth'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        householdId:
          $ref: '#/components/schemas/ResourceId'
          description: Required when scope is household; forbidden for personal.
    PlanningIncomeSourceCreateRequest:
      type: object
      additionalProperties: false
      required: [amount, source, dayOfMonth]
      properties:
        id:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/DecimalString'
        source:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 500
        dayOfMonth:
          type: integer
          minimum: 1
          maximum: 31
    PlanningIncomeSourceUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        amount:
          $ref: '#/components/schemas/DecimalString'
        source:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: [string, 'null']
          maxLength: 500
        dayOfMonth:
          type: integer
          minimum: 1
          maximum: 31
        version:
          type: integer
          minimum: 1
    PlanningAllocationCreateRequest:
      type: object
      additionalProperties: false
      required: [targetType, targetId, allocationMode, allocationValue]
      properties:
        id:
          type: string
          format: uuid
        targetType:
          $ref: '#/components/schemas/AllocationTargetType'
        targetId:
          $ref: '#/components/schemas/ResourceId'
        comment:
          type: string
          maxLength: 500
        allocationMode:
          $ref: '#/components/schemas/AllocationMode'
        allocationValue:
          $ref: '#/components/schemas/DecimalString'
        recurrenceType:
          $ref: '#/components/schemas/AllocationRecurrenceType'
          default: regular
        isSavingsGoal:
          type: boolean
          default: false
          description: When true, targetType must be investment_asset_category.
        goalTargetAmount:
          anyOf:
            - $ref: '#/components/schemas/DecimalString'
            - type: 'null'
        goalDueMonth:
          anyOf:
            - $ref: '#/components/schemas/PlanMonth'
            - type: 'null'
    PlanningAllocationUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        targetType:
          $ref: '#/components/schemas/AllocationTargetType'
        targetId:
          $ref: '#/components/schemas/ResourceId'
        comment:
          type: [string, 'null']
          maxLength: 500
        allocationMode:
          $ref: '#/components/schemas/AllocationMode'
        allocationValue:
          $ref: '#/components/schemas/DecimalString'
        recurrenceType:
          $ref: '#/components/schemas/AllocationRecurrenceType'
        isSavingsGoal:
          type: boolean
          description: When true, targetType must be investment_asset_category.
        goalTargetAmount:
          anyOf:
            - $ref: '#/components/schemas/DecimalString'
            - type: 'null'
        goalDueMonth:
          anyOf:
            - $ref: '#/components/schemas/PlanMonth'
            - type: 'null'
        version:
          type: integer
          minimum: 1
    PlanningPlanCopyRequest:
      type: object
      additionalProperties: false
      required: [targetMonth]
      properties:
        targetMonth:
          $ref: '#/components/schemas/PlanMonth'
...
    PlanningPlanEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PlanningPlanDto'
    PlanningPlanSummaryListEnvelope:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlanningPlanSummaryDto'
    PlanningIncomeSourceEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PlanningIncomeSourceDto'
    PlanningAllocationEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PlanningAllocationDto'
...